Save results from jobs ran on celery workers
