+++ api/tasks.py
@@ -14,7 +14,7 @@ def add(self, x, y):
     return x + y
 
 @app.task
-def run_image(image, command=None, **kwargs):
+def run_image(jobId, image, command=None, **kwargs):
     client = docker.from_env()
     container_output = client.containers.run(image, command=command, **kwargs)
     return container_output
