+++ docker-compose.yaml
@@ -68,9 +68,11 @@ services:
       - REDIS_URL=redis://redis
       - SITE_NAME=SITE_1
       - RABBITMQ_URL=pyamqp://guest:guest@rabbit//
+      - RESULTS_API=http://api:8000/api/v1/results/
     links:
       - rabbit
       - redis
+      - api
     volumes:
       - /var/run/docker.sock:/var/run/docker.sock
     depends_on:
@@ -88,6 +90,7 @@ services:
     links:
       - rabbit
       - redis
+      - api
     volumes:
       - /var/run/docker.sock:/var/run/docker.sock
     depends_on:
